Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-animations/keyframes-remove-documentElement-crash-expected.txt

CONSOLE MESSAGE: line 99: TypeError: null is not an object (evaluating 'document.body.appendChild')
#PID UNRESPONSIVE - WebKitTestRunner (pid 84570)
FAIL: Timed out waiting for notifyDone to be called
